sztt/requirements.txt
amqp==2.4.1
anyjson==0.3.3
autopep8==1.4.3
beautifulsoup4==4.7.1
billiard==3.6.0.0
certifi==2018.11.29
chardet==3.0.4
Django==2.1.7
django-celery==3.2.2
celery==4.3.0rc1
django-cors-headers==2.4.0
djangorestframework==3.9.1
gunicorn==19.9.0
idna==2.8
kombu==4.3.0
pycodestyle==2.5.0
PyMySQL==0.9.3
pytz==2018.9
redis==3.2.0
requests==2.21.0
soupsieve==1.8
urllib3==1.24.1
lxml==4.3.1
vine==1.2.0